requirements.txt
appdirs==1.4.0
Django==1.8.2
django-admin-bootstrapped==2.5.7
django-bootstrap3==8.2.2
django-embed-video==1.1.0
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.6.2
flake8==3.3.0
jedi==0.10.2
mccabe==0.6.1
mock==2.0.0
mongoengine==0.13.0
olefile==0.44
packaging==16.8
pbr==1.10.0
Pillow==4.1.1
psycopg2==2.6.2
pycodestyle==2.3.1
pyflakes==1.5.0
pymongo==3.4.0
pyparsing==2.1.10
requests==2.12.3
six==1.10.0
stripe==1.43.0
